+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - TableBuilder supports multi-line content
  - Added eIcdFileMode for IO platform agnosticism
 
+### Changed
+ - Rewrote JsonItemWrapper serialization for JsonConvert friendliness
+
 ## [11.0.0] - 2020-03-20
 ### Added
  - Added Not null tag for ICDUriBuilder Constructor that takes a URI as an argument.